requirements.txt
fastapi==0.109.0
python-dotenv~=1.0.0
spotipy~=2.23.0
redis~=4.6.0
uvicorn==0.15.0